Rename ethereumInterface -> contractInterface
